bot/requirements.txt
redis
discord.py
